+++ src/api/mod.rs
@@ -50,6 +50,7 @@ use crate::config::{Auth, Config};
 use crate::constants::{ARCH, DEFAULT_URL, EXT, PLATFORM, RELEASE_REGISTRY_LATEST_URL, VERSION};
 use crate::utils::file_upload::LegacyUploadContext;
 use crate::utils::http::{self, is_absolute_url};
+use crate::utils::non_empty::NonEmptySlice;
 use crate::utils::progress::{ProgressBar, ProgressBarMode};
 use crate::utils::retry::{get_default_backoff, DurationAsMilliseconds as _};
 use crate::utils::sourcemaps::get_sourcemap_reference_from_headers;
@@ -991,6 +992,7 @@ impl<'a> AuthenticatedApi<'a> {
             .convert_rnf(ApiErrorKind::ProjectNotFound)
     }
 
+    #[deprecated = "release bundle uploads are deprecated in favor of artifact bundle uploads"]
     pub fn assemble_release_artifacts(
         &self,
         org: &str,
@@ -1019,7 +1021,7 @@ impl<'a> AuthenticatedApi<'a> {
     pub fn assemble_artifact_bundle(
         &self,
         org: &str,
-        projects: &[String],
+        projects: NonEmptySlice<'_, String>,
         checksum: Digest,
         chunks: &[Digest],
         version: Option<&str>,
@@ -1031,7 +1033,7 @@ impl<'a> AuthenticatedApi<'a> {
             .with_json_body(&ChunkedArtifactRequest {
                 checksum,
                 chunks,
-                projects,
+                projects: projects.into(),
                 version,
                 dist,
             })?
@@ -1468,7 +1470,7 @@ pub struct FetchEventsOptions<'a> {
     pub sort: &'a str,
 }
 
-impl<'a> FetchEventsOptions<'a> {
+impl FetchEventsOptions<'_> {
     /// Generate query parameters as a vector of strings
     pub fn to_query_params(&self) -> Vec<String> {
         let mut params = vec![format!("dataset={}", QueryArg(self.dataset.as_str()))];