+++ src/commands/files/upload.rs
@@ -153,7 +153,7 @@ pub fn execute(matches: &ArgMatches) -> Result<()> {
 
     let context = &UploadContext {
         org: &org,
-        projects: &project.into_iter().collect::<Vec<_>>(),
+        projects: project.as_slice().try_into().ok(),
         release: Some(&release),
         dist,
         note: None,